COZMO-5425 Add camera image annotations to SDK (#2077)
* COZMO-5425 Add camera image annotations to SDK

Adds an annotate module which allows for customizable annotations and
updates the tkview module to display annotated images, as well as
locking the aspect ratio for display and scaling the image as the window
size changes.

At some point we need to revisit the tk viewer as it's pretty hacky
right now.

* incorporate pr feedback

* Update Pillow and NumPy required package versions

* fix docstring

* fix typo

* Cleanup handling of missing imports for SDK camera

If the NumPy or PIL packages are not available, ensure the module
remains usable, but methods/functions that require those libraries raise
an exception.

Update the tkrunner so that it exits cleanly in such a scenario - eg. if
tkinter is installed, but numpy is not.