COZMO-5601 COZMO-5676 COZMO-3910 Audio crash fix (#2202)
Updates the `lib/audio` pointer and sets the appropriate flag
on `AudioCallbackContext` objects we create in the Cozmo repo to
execute callbacks on the main thread instead of asynchronously.

lib/audio ce7b73d...c80dd95 (1):
  > Merge pull request #25 from anki/synchronous-audio-callbacks

Description of the audio changes, from that commit message:

When audio callbacks come in from the wwise thread, they're
currently executed on that thread while referencing objects owned
by the main thread. The primary method of thread synchronization
between the wwise thread and the main thread (to make sure those
objects still exist) is for objects to have a `shared_ptr` that
`weak_ptr`s stored in the callback can reference to make sure
the object is still alive before it starts, but unfortunately
that doesn't do anything to protect against the possibility that
the main thread deletes the referenced object _while_ the callback
is running on the wwise thread.

This change seeks to increase stability by adding the option for
client projects (since this is a project-agnostic repo) to specify
that callbacks should not run asynchronously. If this option is
specified, callback data will be put into a queue to be pulled
from in `AudioEngineController::Update` (which is presumably
executed from the main thread), where the main thread will be
unable to delete pertinent objects during callback execution.
As long as the thread that calls `::Update` is the same thread
responsible for deleting objects associated with audio callbacks,
the `weak_ptr` check mentioned above then becomes a guarantee that
those objects will not be deleted during the execution of the
callback.

The stability of this change does depend on the assumption that
the main thread is the one responsible for deleting objects that
post audio callbacks, and since I don't know the audio code super
well someone who's more familiar should verify that that is indeed
the case.

The default behavior with this change alone is the same as before,
but with small changes in the cozmo repo (adding a call to
`SetExecuteAsync(false)` when creating new `AudioCallbackContext`
objects) it will use the new behavior. In a brief test consisting
of one game of speed tap, nothing bad happened with these changes.