COZMO-5790 Cozmo sometimes does fast 360 turns (#2238)
* COZMO-5790 Cozmo sometimes does fast 360 turns

The issue was with paths that contained a backwards segment then forwards segment or vice versa. When we applied the motion profile to the path there was an assumption that we could not change the sign of our speed without stopping first (a point turn). This assumption resulted in the sign of the speed of one of the segments to get flipped.

This is fixed by checking that if the signs of the current segment and the next segment do not match then we make sure to apply the correct sign to the current segment