COZMO-5866 Fixing Meet Cozmo reactionary behavior edge case. (#2276)
In the case where there are no enrolled faces and we enter Meet Cozmo to enroll a face. Reactionary behaviors were not being turned off because it relied on
FaceSlideState to enable/disable it. The fix is to explicitly disable the reactionary behavior in the case of "first time enrollment"