COZMO-967 Stack overflow due to recursively cancelling actions (#2326)
* COZMO-967 Stack overflow due to recursively cancelling actions

One of the causes of the operator delete(void*) crash was due to a stack overflow from recursively deleting/cancelling actions. One of the handlers of RobotCompletedAction was cancelling actions which caused another RobotCompletedAction message to be emitted and the handler would again cancel the action.

This is fixed by keeping a set of the actions that are in the process of being destroyed and if the action is already in the set then don't try to delete it again.