COZMO-6177: Removed pitch drift check because too easily triggered falsely. (#2355)
If we suspect that somehow pitch could drift, but not headingAngle then we can reactivate this, but we need to add a check to make sure that the rate of pitch change is consistent otherwise it's too easy to trigger accidentally from just moving him a little. Currently, I'm assuming the gyroZ check is sufficient.