COZMO-6029. Fix Unknown Error Crash. (#2678)
COZMO-6029. I've verified that this happens on Android when the OBB is not available on device. This could happen if the OBB download failed or, if it was installed to an SDCARD, when the SDCARD is not present. At this point of the initialization DAS is not available so we are using Debug.LogError. The problem is that HockeyApp reports all errors as exceptions. We are downgrading these LogErrors to Logs since the app handles the error correctly and shows the user a message. (#2678)