COZMO-6515 Freeplay taps (#2533)
COZMO-6515 Freeplay Taps

Double tapping a cube while Cozmo is in freeplay will force Cozmo to do something with the tapped cube (pickup, pop a wheelie, stack, or roll). At any point before the "main" part of the behavior other cubes can be double tapped and Cozmo will restart the behavior to use the new cube. If Cozmo does not know where the cube is he will turn towards where he thinks it is and do a quick side to side search for it. If he still can't find it then the tap interaction ends. When a solid blue cube is double tapped it will turn solid green, pulsing blue turns pulsing green. Cubes can not be double tapped while Cozmo is reacting to something.