COZMO-6869 fix freeplay tap animation cut off and other bugs (#2766)
* COZMO-6869 dont cut off the end of tap behaviors

- Not killing tap interaction behaviors as soon as the behavior objective is acheived

- No longer quiting behavior as soon as tapped object is not desired (it becomes unknown or we want to leave the tap interaction goal)

- Fixed warnings about duplicate unlocks (IReactionary behavior was double-subscribed to the message)

- Added code to disable tapped object interaction (and SmartDisable for IBehavior)

- Fixed issue where pop-a-wheelie in freeplay could trigger a double-tap by disabling double-tap while popping a wheelie

* Fixes issue where multiple objects could turn green from double taps

The issue was due to the order functions were being executed in BehaviorManager::Update(). If an object was double tapped, the requested ObjectTapInteraction goal would take an additional tick to become active since the AIGoalEvaluater was Choose[ing]NextBehavior() before the call to UpdateBehaviorTapInteractionObject(). After the tapped object was updated we then call CheckForComputationalSwitch() to try to switch to a reactionary behavior. At this point ReactToDoubleTap is likely runnable and is therefore chosen to start running, the ObjectTapInteraction goal is still not active. The next BehaviorManager::Update(), a reactionary behavior is running (ReactToDoubleTap) so ChooseNextBehavior() is not called which means the AIGoalEvaluator requestedGoal will still not become active. At this point, if another object is double tapped all we will do is update the current double tapped object and request that the goal becomes ObjectTapInteraction, since our previous request still hasn't taken effect. ReactToDouble now incorrectly updates the object lights since there have been two different current tapped objects since it started running.