COZMO-7386 Negative detected quad area (#2828)
* COZMO-7386 Negative detected quad area

ImageIMU data used for calculating rolling shutter pixel shifts could have invalid zero timestamps which results in invalid pixel shifts. Once these invalid shifts are applied to the detected quads it warps the quads enough to cause them to have negative areas. The zero timestamps were caused by imageIMU data coming in after the corresponding image had been processed so the imageIMU data's timestamp would never be calculated.

The zero timestamp issue is fixed by checking the timestamps of the imageIMU data from the previous image when calculating timestamps for the current image and if any of the previous data has a zero timestamp then it is immediately calculated.

Apart from fixing the zero timestamp issue we decided to just use the line number the imageIMU data was captured at to do the linear interpolation instead of timestamps when calculating the pixel shifts