COZMO-7509 Warning: BehaviorManager.LeaveObjectTapInteraction.NullChooser (#2891)
If the current behavior chooser was null or did not support object tap interactions, we would not clear the object tap interaction

The effect of this bug is that we would get stuck with an object tap interaction while in a minigame so the warning would spam as we repeatedly try to leave tap interaction. The good news is that nothing is affected by this code path (bug does not cause anything besides spamming). As soon as the minigame ends we will be able to clear object tap interaction and the warning will stop.