COZMO-7196. Fix SecurityException (#2906)
* COZMO-7196. Commenting out code that was causing a SecurityException on Android 6 and above.

* COZMO-7196 Catch SecurityException from accessing wifi results (#2908)

It seems that on some devices running Android 6.0, but not others,
calling `WifiManager.getScanResults()` throws a `SecurityException`
that we weren't dealing with, causing the app to crash. This change
catches the exception and treats it as if we received a null result
list.

(cherry picked from commit f5a77f4b1a9e99457719ff22d085b7df64002f61)