COZMO-6387: Fixes “Drop Cube” context button from showing up prematurely while Cozmo is still in the middle of the pickup action. (#2998)
The carry state, upon which the button is based, can go true during a pickup action but the action itself can end up failing following other checks to verify if the cube is actually in the lift at which time the carry state is updated to false. When such a failure occurs the “Drop Cube” option appears (though disabled) for a few seconds and then disappears when the action fails. To make things a little cleaner this fix defers the update of the context button until after the pickup action has completed.