COZMO-7866 Android wifi connectivity flow (#2942)
This change adds a new connectivity flow for Android that leverages our ability to initiate a new wifi connection through API calls. Instead of requiring users to go into their OS wifi settings to connect to Cozmo, this change will allow them to select a Cozmo detected via wifi scanning, enter its password, and connect to it without ever leaving the app.

Some complications are addressed to make this happen. On newer android versions (some devices with 6, most/all devices with 7) the OS will not communicate over a network that doesn't provide internet, like Cozmo; we have to track connection state and make an OS call to bind our sockets to the Cozmo network in order to communicate over it. Additionally, on Android 6+ we need to request coarse location permission from the user and have the GPS turned on in order to be able to get wifi scan results.

The new flow is added as a group of new Unity prefabs, inserted where the flow would normally kick to the `SearchForCozmoScreen` prefab. Instead of that screen, we go to a group of screens that walks the user through the connection process; on success, we then push to `SearchForCozmoScreen` where a connection should be detected, and on failure we go to `SearchForCozmoFailedScreen`.