COZMO-9038 Fix erase pattern in MultiClientComms::Update (#3051)
Found while looking at code; confirmed with KevinY/Raul/Brad. Technically this is not a bug but a slightly dangerous pattern. In two cases in this file, a loop was iterating a map and conditionally erasing items. On the erase, the call was erase(it++);. Changing to the pattern of it = erase(it), which is now allowed for maps in C+11.