COZMO-9054 COZMO-9058 Start ping test before connecting (#3063)
* COZMO-9054 COZMO-9058 Start ping test before connecting

When the Android connection flow is entered, it starts regularly
pinging Cozmo to detect when we've successfully connected.
However, because these pings only start when we enter the flow, we
can't know if we're already connected to Cozmo when starting until
we get a result back from the ping test. This means if you're
already connected when beginning the flow, you'll still catch a
brief glimpse of one of the screens inside it before continuing
connection, which is ugly.

This change directs our Android build to start regularly pinging
in Java as soon as we enter `IntroManager`'s flow. That way, if
we're already connected when we start the connection process,
we'll know right away and we can just bypass the special Android
connection flow entirely.

* Only start ping test if we plan to use special flow