COZMO-7870 Integrate existing search view into Android flow (#3104)
Once we begin connecting to the robot, we can make the Android
flow more seamless with the normal experience by using the
existing "Searching for Cozmo..." screen instead of a special
prefab that does essentially the same thing. To achieve our
Android-specific goal of adding a "cancel" button if our
connection attempt takes too long, this change repurposes the
existing AndroidConnectToNetwork prefab to live as an overlay
on top of the SearchForCozmoScreen view.