COZMO-7741 Block Android loading screen until permission given (#3147)
Without the READ_EXTERNAL_STORAGE permission, the Play Store
version of our app is unable to read from the expansion (.obb)
file, causing app loading to stall permanently.

This change adds a check for whether we have that permission or
not before loading begins. If we don't have the permission, we
instantiate a screen that sits over the loading screen and does
not exit until the user gives us the permission. If the user hits
"don't ask again" while denying the permission, the button that
would normally prompt them for the permission will instead take
them to the App Settings screen, where they can enable it
manually.