COZMO-9013: Fist bump (#3146)
COZMO-9015: Fist bump behavior
COZMO-9042: Fist bump spark
COZMO-9043: Spark UI
COZMO-9044: Spark celebration
COZMO-9045: Fist bump in freeplay

Two jsons define the spark version (sparksFistBump.json) and the freeplay/celebratory version (fistBump.json). The spark version looks a little longer for a face and attempts fist bump regardless of whether or not one was found, while the freeplay version searches for a face for a shorter period of time and only requests bump if a face is found.

The behavior is triggered as a reaction to certain BehaviorObjectives (in reactionTrigger_behavior_map.json) with some cooldown time and some probability of triggering which allows it to run the same way during freeplay as it does when celebrating after a sparking behavior. 

* Removed seemingly redundant BehaviorObjective::BlockStacked. (There’s already a StackedBlock)

* Added E2R::EnableMotorPower to disable head motor in addition to lift motor to eliminate false fist bump detections. Animation seems to sometimes not stop head immediately.

* Added debug var to allow fist bump to always trigger on specified behaviorObjectives for testing purposes. Effectively sets cooldownTime to 0 and triggerProbability to 1.

* Added BehaviorObjective::PerformedStrongWorkout as a trigger for fist bump