COZMO-8533 Cozmo gets stuck during double tap (#3170)
* COZMO-8533 Cozmo gets stuck during double tap

If you double tapped a cube and then put another cube on top of it, Cozmo would get stuck switching between different behaviors. This was due to the bestObjectForAction not being updated while we have a double tapped object (this forces the best object to be the double tapped object). The behavior would start running and then realize it can't use the tapped object so it would end and another behavior would be picked. The new behavior would start and realize it also couldn't use the tapped object and this would continue to cycle.

This is fixed by checking if the tapped object is valid for the ObjectUseIntentions the object tap interaction behaviors use and if it is not valid for any of them then we bail out of the tap interaction goal.