COZMO-9259 Docking with placement offsets should line up (#3154)
* COZMO-9259 Docking with placement offsets should line up

When doing a DriveToPlaceRelObjectAction with placement offsets, we were not modifying the preDock pose to drive to. This resulted in weird maneuvers being made by the dock action attempting to place at the offsets.

The DriveTo action now adjusts the preAction poses it is going to drive to by applying the placement offsets to them.