COZMO-7855 Fix new flow instances blowing away ping test (#3243)
Since Unity has no lifecycle method that runs synchronously with
a GameObject being destroyed, calls in AndroidConnectionFlow's
`OnDestroy` method to stop the Java ping test from running were
taking place _after_ a new instance had spun up and (redundantly)
attempted to start one of its own, instead of the other way around.
I've attempted to address this before by introducing the `Disable`
method in AndroidConnectionFlow, but failed to call it in all the
places where we destroy existing flow instances.

This change makes a new method to destroy the ACF instance (with
a call to Disable) and has all existing locations that destroy
ACF use the new method instead. This should ensure that the ping
test is not inadvertently stopped right after a new instance of
the flow is created.