COZMO-9445 Fix search failed screen crashing w/ no ping status (#3276)
When initializing the SearchFailed screen from the Android flow
(in the situation where we don't find any Cozmos through scanning),
we don't initialize it with a PingStatus object because the Android
flow is detecting connections through Java instead. This caused a
crash because `_PingStatus` is assumed to not be null.

This change checks if `_PingStatus` is null before using it.