Stop PR from building if there is a merge conflict. (#3271)
BI-235: Builds will not trigger if there is a merge conflict on the PR
via github. However if a build was kicked of manually by the developer
or the build is in queue when a merge conflict is created a weird thing
will happen. An older version of the PR will be built. Leading to false
positives and false negatives.

Github only updates the */merge branch is a merge is possible after the
latest commit to the PR or master.  This script will compare the time
stamp of the PR against the merge.  As a clean merge branch will always
be newer.  If this is not the cause.  The build will error out prevent
the stale build from happening.