COZMO-8856 fixing unclose-able low battery dialog bug (#3287)
I couldn't reproduce this, but I think it has to do with the low battery dialog being driven by an update loop that is depending on _LowBatteryAlertTriggered to be set in order not to call OpenLowBatteryAlert once per frame. Since that flag was in a completion callback before it could have been spawning a ton of dialogs.