COZMO-9328 Fix cube light flicker when stopping/starting a new light … (#3255)
* COZMO-9328 Fix cube light flicker when stopping/starting a new light anim

When we stop animations, update() is called and that will pick a next default anim should there be no animations left to play. When this happens the lights immediately get set for the default anim. This means if you call Stop and then Play, you may see a brief flicker as the lights get set for the default anim and them immediately get set to the animation you want to play.

This is fixed by adding a specific function that handles stopping an animation and playing another. It internally pass a flag through update to prevent the picking of a default anim