COZMO-8004 Cozmo refuses to enroll a face (#3289)
* COZMO-8004 Cozmo refuses to enroll a face

This is caused by the faceRecognizer thread not running. In order for the thread to run _isRunningAsync needed to remain true. However, just before the thread was created the isRunningAsync flag was set to false. Immeditately after the thread was created was the flag set to true. Depending on when the new thread executed the Run function _isRunningAsync could be either true or false.

This is fixed by setting the flag before the thread is created