COZMO-9571 Add CAPTIVE_PORTAL_CHECK to list of connected states (#3322)
Gary found an example of a phone (Note 3 on Android 5) that, after
entering VERIFYING_POOR_LINK state and binding to the network,
entered CAPTIVE_PORTAL_CHECK state and unbound itself because we
don't recognize C_P_C as a "connected" state. This change adds
C_P_C to the list of connected states, so that our code recognizes
a continuous connection in this situation.