COZMO-9587 CustomObject webots test failing (#3364)
We were incorrectly using objectID as a measure of how many objects existed in the world at a given point in time. An object was sometimes being added to the world which would take an id and then immediately get deleted and replaced with a different object. It is unclear why the object was being added in the first place.