COZMO-9761 Android - Add description for Java crashes
When HockeyApp reports a Java crash we want to include a description
with the `apprun` and `deviceId`.  Unfortunately, we were using the
`startHockeyAppManager` method in the `HockeyUnityPlugin` which does NOT
allow us to specify either a crash description or supply a
`CrashManagerListener` that can return one.

This change adds a new method to `CozmoActivity` called
`startHockeyAppManager` that does allow us to supply a crash
description.  It is implemented based on the HockeyUnityPlugin 1.1.2
code as a reference.