COZMO-9692 Fix null WifiInfo dereference (#3421)
My last attempted fix attacked this from the wrong angle - the
NetworkInfo was valid, but the WifiInfo wasn't. According to
Android documentation, that can happen if the wifi is not in a
"CONNECTED" state, but apparently most devices provide a WifiInfo
anyway.

This fix makes sure the WifiInfo isn't null before we attempt to
use it. Unfortunately, the reason we're attempting to use it here
is because we're seeing if a network that you're _almost_
connected to (in VERIFYING_POOR_LINK state, for example) is a
Cozmo network so we should bind to it, but unfortunately we can't
check that without a WifiInfo, so this would delay binding to the
Cozmo network (and thus being able to communicate with the robot,
if the device wants to restrict our communication) until the
device actually gives us one.