COZMO-9315 Add option to turn backpack light off
Add an option to turn the backpack light off. This is one of the 9 options from the dropdown of backpack color options. With this change, the option to turn the backpack light off is the black lightbulb (which has a different art style than the other lightbulbs, and this will be fixed when we integrate the icons from Seth). The black option replaces the magenta option (and a purple option still remains).