COZMO-9907 Cubes and robot in crazy poses (#3500)
* COZMO-9907 Cubes and robot in crazy poses

We think Cozmo was localizing to a cube while the cube was moving. This could be caused by moved messages being ignored while waiting for a double tap.

We will now refuse to localize to an object that has been recently tapped (the time while we are waiting for a double tap currently 500ms)
We will also mark the cube as dirty when the "ignoring movement while waiting for double tap" timeout is reached unless we get another tap.
Cubes will now be marked as dirty when we receive the stop message following a moved message (instead of just marking dirty on the moved message).
Marking a cube as anything but known will now automatically cause the robot to become unlocalized from that object.