COZMO-10147 android.py - Add support for NDK r14
NDK r14 is the latest version from Google.  These changes add support
for downloading and installing NDK r14 when building Cozmo.

Unfortunately, the current version of buck that most of us have
installed does NOT recognize NDK versions past 13.  As a result, if you
try to build a buck target it will fail if the value of ndk_version in
the ndk section of .buckconfig is greater than 13.  To workaround this,
I updated the tools/ankibuild/android.py script to look for the NDK
version first in .buckconfig but then also in .buckconfig.override.
Later, when a new version of a buck is released, we can update
.buckconfig and remove the .buckconfig.override file.