COZMO-9302 Scratch Blocks: get iOS keyboard and sound working (#3598)
Summary: Fix keyboard and sound on iOS 9 by using UIWebView. Fix sound on iOS 10.

Fixes two bugs on iOS 9 by using UIWebView instead of WKWebView. The two bugs are:

When try to change text on the “Cozmo Says” block, keyboard would slide up and then immediately slide off, without any user interaction.
Scratch sounds to connect and trash the blocks did not play.
One of the benefits of WKWebView is improved performance, but the difference is not evident with all content, and I don’t see a difference with Scratch.

Also fixes Scratch blocks sound for WKWebView (which we now use on iOS 10+ only).