COZMO-9405 Disconnect on Android if Cozmo already in use (#3622)
If connecting to a Cozmo that another device is already using,
hitting "cancel" will restart the connection process, but because
the Android flow detects that you're already connected to a Cozmo
it skips through the part where you would attempt to choose a new
Cozmo and tries to finish the connection. The solution to this is
to disconnect from Cozmo on Android if the user cancels out of the
"this Cozmo is already in use" screen.