COZMO-9417 Scratch Blocks close button (#3626)
Add close button to upper left corner of webview to close out of Code Lab and return to minigame menu. Will work with Seth to get a new close button asset in svg with up and down states.

Updated code to not disconnect from robot when exiting CodeLab, and to return to freeplay mode.

Leave sound on while in Code Lab.

Also removed some unnecessary Scratch images and code.