COZMO-10130 Multiple cubes are solid green when Cozmo is stacking (#3634)
If a cube was double tapped while PickupBlock was running, it would transition us into StackBlocks. Sometimes this would cause both cubes on the ground to turn green. This is due to a variable not being reset properly during StackBlocks init. It would lead to the closer of the two blocks to be selected by StackBlocks regardless of which on was double tapped. So if you double tapped the farther block it would turn green from being double tapped but the behavior would select the closer one and the DriveToObjectAction would turn that one green.

Also fixes a problem with ReactToDoubleTap being able to run while Cozmo is putting a block down (not considered pickingAndPlacing since an animation is putting the block down)