COZMO-10286 Random crashes on PR Android builds (#3644)
We think the crash was caused by a bad cast going into a switch. A uint was being cast to an enum and then we were switching over the enum with no default case. None of the code in the switch will be executed so the function has no return which is undefined