Prevent Webots from crashing when reloading a world (#3753)
* Prevent Webots from crashing when reloading a world by making Init() in physics controllers properly close socket.
For some reason, since 52db9e949554e543ed7aa4cd5776253519a36a09, reloading a world in Webots does not fully reset the cozmo_physics plugin. The result is that the UDPServer in physVizController is retaining state and so when it receives the 1 byte “hello” message from a client it thinks it already knows about it tries to interpret it as a CLAD message and crashes.

Thanks Baustin!