COMZO-10429 Mem map use new clad messages (#3794)
* COMZO-10429 Men map use new clad messages

The gist is to make the debug visualization of memory map (both in webots, and in the app
in the debug console) use the new memory map clad messages created for sending
memory map data to the SDK.
This is a WIP, however the scheme now works for the webots viz.  The old scheme is still
in place, so currently in webots you’ll see two identical sets of quads drawn.
In the old scheme, the sending code would do a DFS on the memory map quad tree, and
create ‘simple quads’, and send them, and the receiving code would just render those
quads.  In the new scheme, the sending code does a DFS on the tree and sends ‘quad
Info’ which is more compact.  Then on the receiving end, the data is decoded and the
tree effectively reproduced, along with simple quads to draw.
Still to do:
1. Set z height of the tree based on which tree it is
2. Implement the C# receiver/decoder for use in the debug console memory map viz.
3. Remove the old sending code
4. Attempt to unify the ‘external’ and ‘internal’ messages/schemes, to reduce code.
  (The issue is that the SDK quad info is simpler and different from the internal quad
  Info.)

* COZMO-10429 continued

- Set z position of quad tree rendering as we did before
- Support deletion of tree renders:  Had to put unique ID (lower half of pointer to quad tree)
  in the memory map info structure, so that the engine could call the viz to ask it to be
  deleted.
- Some perf measurements

* COZMO-10426 memory map new clad messages continued

- Change ID-ing by putting the ID in on the engine side, and sending as a string in
  the ‘begin’ message
- Fix Release link error by including point impl
- Disable the new system, and add comments in code explaining why (we can’t enable
  the new system until it is implemented in the C# visualizer; otherwise we’ll have two
  sets of quads rendering in webots)

* COZMO-10429 memory map use new clad messages

- Remove some scaffolding and spam

* Auto committing firmware from wip/PaulT/COZMO-100429-Mem-map-use-new-clad

* COZMO-10429 Memory Map use new clad structures

This is the final phase:  Make the Unity/C# debug visualizer use the new system, and
deactivate the old system.

- Gut the old ‘Draw’ method so it just does the processor drawing (but still call it)
- Remove other code that was doing the old scheme
- In BroadcastInternal, respect and reset the _gfxDirty and kRenderNavMeshQuadTree
  flags
- In the ‘end’ handler for the new scheme, remove the data that is no longer used
- In MemoryMapNode, remove the unused _content member
- Implement the new scheme in Unity/C#:  Message handlers, decoding, etc.

* COZMO-10429 memory map use new clad

- PR feedback (mostly renames)
- Also fix il2cpp iOS build by updating the IOS il2cpp project

* COZMO-10429

Needed updated IOS project for il2cpp