COZMO-10566: Fixed problem where commanding DriveStraightAction backw… (#3851)
* COZMO-10566: Fixed problem where commanding DriveStraightAction backwards travelled further than commanding the same distance forwards.
This was because the default acceleration rate is slower than the deceleration rate and slowing down while going backwards uses acceleration. (We may want to change this as you could argue it doesn’t make intuitive sense. COZMO-10743)

Fix is to make it stop immediately when path is complete. Since path follower slows down to a nominal speed of 20mm/s, it should brake quickly once it’s drive center has reached the end of the path.

* Auto committing firmware from kevin/COZMO-10566-Fwd_and_Back_unequal_distances