COZMO-2827: Refactor AnimationStreamer to create abstract interface (#3871)
Extract public interface points from AnimationStreamer to declare new
abstract IAnimationStreamer.  This interface identifies the methods
that must be supported by Animation Streamer 2.0 so it can be used as a
drop-in replacement without disruption.

I left the new class in the same file (AnimationStreamer.h) to simplify
review.  I don’t see an immediate need to create a separate file.

Some of the AnimationStreamer methods could be made protected or
private because they are not being used outside the class.  I decided
to make this change at this time, to simplify review.

Meta-goal here is to make incremental progress toward deployment of
Animation Streamer 2.0.  I’d like to avoid doing all the changes at
once.