COZMO-10630 Change splash screen scale mode to fix crash (#3882)
Digging into this crash, in which the app crashes on bootup on
older devices, it appears that:
- the callstack for the crash includes a Unity class method named
  `onLayoutChanged`
- looking through their source code, the only class containing a
  method by that name is `StaticSplashscreen`
- the only place `StaticSplashscreen` is instantiated is in the
  initialization of `UnityPlayer`, where it reads the file
  `assets/bin/Data/settings.xml` from the apk and uses the
  "showSplash" setting to determine if the splash screen should be
  used
- it passes into the splash screen the value of the "splash_mode"
  setting from the XML file
- for some reason, on my local builds the value of "splash_mode"
  is 0, but the value of "splash_mode" on build server builds is 2
- my local builds do not crash on these older devices, but build
  server builds do

Therefore: it seems like a good idea to force "splash_mode" to 0
if possible. I haven't the slightest idea why my local builds are
putting 0 in the XML file already when the `ProjectSettings.asset`
file says it's 2, and I haven't the slightest idea why _any_ of
our builds are putting "showSplash = True" in the XML file when
the `ProjectSettings.asset` file says we're not using the splash
screen, but maybe this change will help build server builds do the
right thing.