COZMO-9287 Code Lab categories (#3910)
COZMO-9287 As a Scratch user, I would like to see a limited set of blocks at a time (categories)
COZMO-10718 Break out animation block into 15 separate blocks in new pink "actions" category.

Also modified default_toolbox.js and horizontal_playground.html to refer to cozmo code.

vm* and dist.js are generated files.

Still to do, for later:

In this PR I temporarily turned on the toolbox scrollbar in order to workaround a bug in scratch-blocks (waiting for them to fix). The toolbox scroll bar will currently always show regardless of number of blocks in the toolbox. If necessary we can try to fix the bug as well.
Font for categories not yet set to desired font. Need to set to Avenir but currently is Helvetica.