COZMO-10827: Clamp Poses to Flat (#3916)
Clamps poses of all _active_ observed objects to flat if they are already within the tolerance for localization.

Address this problem:
1. Let see Cozmo see cube from one side and localize to it
2. Pick up Cozmo and place him on the opposite side of the cube
3. Observe that Cozmo will often appear to sink into the floor and/or his world model will be severely tilted
4. Repeating can make the problem get worse and worse.

The issue is that the estimated rotation of the cube's pose (around robot's Y axis) is often tilted a few degrees, but still within the IsRestingFlat threshold of 5º for localization. So Cozmo localizes to the tilted cube. When picked up and placed on the other side of the cube, he also mis-estimates the cube's tilt by the same few degrees, but _still in the same direction relative to the his position_. Since the cube is still localizable, Cozmo assumes it has not moved and he has, but the only way to explain the tilt is then to put himself in the floor. So error is effectively additive. The fix is to clamp poses to actually be flat if we are effectively assuming they are anyway.

Other changes related to implementing the above:
* Add GetAngleAroundAxis method to Rotation3d
* Add helper to convert AxisName enum to the corresponding Vec3f
* Exposing angle tolerance for what to consider flat (as virtual method)
* Add non-templated AxisToIndex method
* Return rotated parent axis in GetAngularDeviationFromParentAxis as optional output arg
* Define pure virtual IsActive in Vision::ObservableObject
* Cleanup: remove ancient “#if 0” block in ObservableObjectLibrary
* Use Quad2f::Contains check in finding matching marker for updating observed times, instead of 5-pixel centroid distance threshold (more lenient in case of clamping)