COZMO-10247: Don't retain state of unreacted targets (#3921)
Reaction trigger could misfire if robot sees a pet while off treads,
because reaction trigger retains petID as target even if robot does not
react.  Once robot is returned to treads, behavior is triggered again
but petID is no longer valid.

Fix by changing trigger strategy to consider active pets only. It needs
to retain list of pets it has already reacted to, but it does not need
to retain list of pets it has NOT reacted to.