COZMO-10853 Sometimes Cozmo is not cured by placing him on his back (#3926)
If hiccups are disabled when the OnTreadStateChanged message is sent then the message is ignored. This was an issue with ReactToCliff. Sometimes ReactToCliff triggers when picking Cozmo up, if you then put Cozmo on his back while he is reacting to cliff that will not be counted as curing hiccups because the OnTreadStateChanged message is ignored since hiccups are disabled during ReactToCliff.

This is fixed by not disabling hiccups during ReacToCliff. Hiccups already can't interrupt other triggered behaviors so disabling is not actually neccessary. Also even if Cozmo does hiccup while reacting to cliff, the hiccup animations should move him back enough so he is no longer seeing the cliff