COZMO-10788 RobotStateHistory.AddRawOdomState.TimestampDeltaTooLarge (#3932)
Was expecting this error to only occur when there was a state message with a garbage timestamp. Turns out it is possible that the connection to the robot is just overloaded so a bunch of the unreliably sent state messages are dropped (some reliable messages still being received so the connection is not dropped). If more than 5 seconds worth of state messages are dropped in such a way then this error would print for all new state messages (they have timestamps >5 seconds since the last thing in history).

This spamming is fixed by only printing the error once per series of consecutive state message with large timestamp deltas. If we get more than a certain number of consecutive state messages with large timestamp deltas then clear history because the state messages must be correct and our history is just old